// frontend/src/types/generated/DataSourceTableInfo.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { DataSourceColumnInfo } from "./DataSourceColumnInfo";
import type { DataSourceForeignKey } from "./DataSourceForeignKey";
import type { DataSourcePrimaryKey } from "./DataSourcePrimaryKey";

/**
 * Information about a single table
 */
export type DataSourceTableInfo = { name: string, columns: { [key in string]: DataSourceColumnInfo }, primaryKeys: Array<DataSourcePrimaryKey>, foreignKeys: Array<DataSourceForeignKey>, };