// frontend/src/types/generated/ChangeTableCondition.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

/**
 * Condition for matching rows in change tables (serializable)
 */
export type ChangeTableCondition = { "type": "OR", conditions: Array<ChangeTableCondition>, } | { "type": "AND", conditions: Array<ChangeTableCondition>, } | { "type": "column-equals", column: string, value: string, } | { "type": "column-matches", column: string, regex: string, } | { "type": "column-not-empty", column: string, };