// frontend/src/types/generated/DataExtractionBlueprint.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { DataSourceConfig } from "./DataSourceConfig";
import type { TableExtractionConfig } from "./TableExtractionConfig";

export type DataExtractionBlueprint = { sources: Array<DataSourceConfig>, tables: Array<TableExtractionConfig>, };