// frontend/src/types/generated/TransformOperation.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ChangeTableCondition } from "./ChangeTableCondition";
import type { JoinType } from "./JoinType";
import type { TransformSource } from "./TransformSource";

/**
 * Defines how a virtual table derives its rows
 */
export type TransformOperation = { "type": "filter", source: TransformSource, condition: ChangeTableCondition, } | { "type": "join", left: TransformSource, right: TransformSource, join_type: JoinType, 
/**
 * Column pairs to join on: (left_col, right_col)
 */
on: Array<[string, string]>, } | { "type": "union", sources: Array<TransformSource>, };