// frontend/src/types/generated/TransformSource.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

/**
 * Reference to either a real database table or a previously computed virtual table
 */
export type TransformSource = { "type": "table", source_id: string, table_name: string, } | { "type": "virtual-table", virtual_table_id: string, };