// frontend/src/types/generated/ValueExpression.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

/**
 * A value derived from column data. Supports column reference, constant, or template.
 */
export type ValueExpression = { "type": "column", column: string, } | { "type": "constant", value: string, } | { "type": "template", template: string, };