// frontend/src/types/generated/VirtualTableConfig.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { TransformOperation } from "./TransformOperation";

/**
 * A virtual table definition
 */
export type VirtualTableConfig = { id: string, name: string, operation: TransformOperation, };